# web/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CampaignDetail": {
        "additionalProperties": false,
        "properties": {
          "campaign": {
            "$ref": "#/components/schemas/CampaignSummary"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceProjection"
            },
            "maxItems": 1000,
            "title": "Evidence",
            "type": "array"
          },
          "extraction": {
            "$ref": "#/components/schemas/ExtractionProjection"
          },
          "record_sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Record Sha256",
            "type": "string"
          },
          "source_document_id": {
            "minLength": 1,
            "title": "Source Document Id",
            "type": "string"
          },
          "source_title": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Title"
          },
          "source_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Source Url",
            "type": "string"
          },
          "validation_issues": {
            "items": {
              "maxLength": 500,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 100,
            "title": "Validation Issues",
            "type": "array"
          }
        },
        "required": [
          "campaign",
          "source_document_id",
          "source_url",
          "source_title",
          "record_sha256",
          "extraction",
          "evidence",
          "validation_issues"
        ],
        "title": "CampaignDetail",
        "type": "object"
      },
...
      "CampaignListResponse": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "facets": {
            "$ref": "#/components/schemas/CampaignFacets"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CampaignSummary"
            },
            "maxItems": 100,
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "facets",
          "as_of"
        ],
        "title": "CampaignListResponse",
        "type": "object"
      },
...
      "ChatAnswer": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/AnswerCitation"
            },
            "maxItems": 20,
            "title": "Citations",
            "type": "array"
          },
          "insufficient_evidence": {
            "title": "Insufficient Evidence",
            "type": "boolean"
          },
          "plan": {
            "$ref": "#/components/schemas/ChatQueryPlan"
          },
          "warnings": {
            "items": {
              "maxLength": 500,
              "minLength": 1,
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "plan",
          "citations",
          "insufficient_evidence",
          "warnings"
        ],
        "title": "ChatAnswer",
        "type": "object"
      },
...
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "question": {
            "maxLength": 1000,
            "minLength": 2,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
...
      "ComparisonDimension": {
        "enum": [
          "rate",
          "term",
          "fee",
          "reward",
          "eligibility"
        ],
        "title": "ComparisonDimension",
        "type": "string"
      },
...
      "ComparisonRequest": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "campaign_ids": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 2,
            "title": "Campaign Ids",
            "type": "array"
          },
          "dimensions": {
            "items": {
              "$ref": "#/components/schemas/ComparisonDimension"
            },
            "maxItems": 5,
            "minItems": 1,
            "title": "Dimensions",
            "type": "array"
          }
        },
        "required": [
          "campaign_ids",
          "dimensions"
        ],
        "title": "ComparisonRequest",
        "type": "object"
      },
...
      "NotificationListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CampaignChangeEvent"
            },
            "maxItems": 100,
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "NotificationListResponse",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "detail": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "correlation_id",
          "errors"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
      "ProductFamily": {
        "enum": [
          "financing",
          "card",
          "investment",
          "participation_account",
          "other",
          "unknown"
        ],
        "title": "ProductFamily",
        "type": "string"
      },
...
      "SalesChannel": {
        "enum": [
          "all",
          "digital",
          "mobile",
          "web",
          "branch",
          "unknown"
        ],
        "title": "SalesChannel",
        "type": "string"
      },
...
      "ValidityFilter": {
        "enum": [
          "active",
          "upcoming",
          "expired",
          "unknown"
        ],
        "title": "ValidityFilter",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Kanıt temelli katılım bankacılığı kampanya analiz API'si.",
    "title": "Katılım Analiz",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/campaigns": {
      "get": {
        "operationId": "list_campaigns_api_v1_campaigns_get",
        "parameters": [
          {
            "in": "query",
            "name": "bank_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bank Id"
            }
          },
          {
            "in": "query",
            "name": "product_family",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProductFamily"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Family"
            }
          },
          {
            "in": "query",
            "name": "currency",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Z]{3}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Currency"
            }
          },
          {
            "in": "query",
            "name": "customer_segment",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Segment"
            }
          },
          {
            "in": "query",
            "name": "sales_channel",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SalesChannel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sales Channel"
            }
          },
          {
            "in": "query",
            "name": "validity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ValidityFilter"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Validity"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 500,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Geçersiz istek"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Kaynak bulunamadı"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sözleşme doğrulama hatası"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sunucu hatası"
          }
        },
        "summary": "List Campaigns",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaign_id}": {
      "get": {
        "operationId": "get_campaign_api_v1_campaigns__campaign_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "maxLength": 191,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9:._-]{0,190}$",
              "title": "Campaign Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Geçersiz istek"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Kaynak bulunamadı"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sözleşme doğrulama hatası"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sunucu hatası"
          }
        },
        "summary": "Get Campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/chat": {
      "post": {
        "operationId": "chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatAnswer"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Geçersiz istek"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Kaynak bulunamadı"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sözleşme doğrulama hatası"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sunucu hatası"
          }
        },
        "summary": "Chat",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/comparisons": {
      "post": {
        "operationId": "comparisons_api_v1_comparisons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparisonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Geçersiz istek"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Kaynak bulunamadı"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sözleşme doğrulama hatası"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sunucu hatası"
          }
        },
        "summary": "Comparisons",
        "tags": [
          "comparisons"
        ]
      }
    },
    "/api/v1/coverage": {
      "get": {
        "operationId": "coverage_api_v1_coverage_get",
        "parameters": [
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CoverageEntry"
                  },
                  "title": "Response Coverage Api V1 Coverage Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Geçersiz istek"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Kaynak bulunamadı"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sözleşme doğrulama hatası"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sunucu hatası"
          }
        },
        "summary": "Coverage",
        "tags": [
          "coverage"
        ]
      }
    },
    "/api/v1/notifications": {
      "get": {
        "operationId": "list_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 500,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Geçersiz istek"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Kaynak bulunamadı"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sözleşme doğrulama hatası"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sunucu hatası"
          }
        },
        "summary": "Kampanya değişiklik bildirimlerini salt okunur listele",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/previews/extractions": {
      "post": {
        "operationId": "preview_extraction_api_v1_previews_extractions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractionPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Geçersiz istek"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Kaynak bulunamadı"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sözleşme doğrulama hatası"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Sunucu hatası"
          }
        },
        "summary": "Yapıştırılan metinden doğrulanmamış çıkarım önizlemesi üret",
        "tags": [
          "previews"
        ]
      }
    },
...
  }
}